--- .vercel/output/_functions/manifest_BP1gu5Kl.mjs
@@ -1,99 +1,0 @@
-import { N as NOOP_MIDDLEWARE_FN } from './chunks/astro-designed-error-pages_DG1WDV-e.mjs';
-import { f as decodeKey } from './chunks/astro/server_T2yNNXBH.mjs';
-import 'clsx';
-
-function sanitizeParams(params) {
-  return Object.fromEntries(
-    Object.entries(params).map(([key, value]) => {
-      if (typeof value === "string") {
-        return [key, value.normalize().replace(/#/g, "%23").replace(/\?/g, "%3F")];
-      }
-      return [key, value];
-    })
-  );
-}
-function getParameter(part, params) {
-  if (part.spread) {
-    return params[part.content.slice(3)] || "";
-  }
-  if (part.dynamic) {
-    if (!params[part.content]) {
-      throw new TypeError(`Missing parameter: ${part.content}`);
-    }
-    return params[part.content];
-  }
-  return part.content.normalize().replace(/\?/g, "%3F").replace(/#/g, "%23").replace(/%5B/g, "[").replace(/%5D/g, "]");
-}
-function getSegment(segment, params) {
-  const segmentPath = segment.map((part) => getParameter(part, params)).join("");
-  return segmentPath ? "/" + segmentPath : "";
-}
-function getRouteGenerator(segments, addTrailingSlash) {
-  return (params) => {
-    const sanitizedParams = sanitizeParams(params);
-    let trailing = "";
-    if (addTrailingSlash === "always" && segments.length) {
-      trailing = "/";
-    }
-    const path = segments.map((segment) => getSegment(segment, sanitizedParams)).join("") + trailing;
-    return path || "/";
-  };
-}
-
-function deserializeRouteData(rawRouteData) {
-  return {
-    route: rawRouteData.route,
-    type: rawRouteData.type,
-    pattern: new RegExp(rawRouteData.pattern),
-    params: rawRouteData.params,
-    component: rawRouteData.component,
-    generate: getRouteGenerator(rawRouteData.segments, rawRouteData._meta.trailingSlash),
-    pathname: rawRouteData.pathname || undefined,
-    segments: rawRouteData.segments,
-    prerender: rawRouteData.prerender,
-    redirect: rawRouteData.redirect,
-    redirectRoute: rawRouteData.redirectRoute ? deserializeRouteData(rawRouteData.redirectRoute) : undefined,
-    fallbackRoutes: rawRouteData.fallbackRoutes.map((fallback) => {
-      return deserializeRouteData(fallback);
-    }),
-    isIndex: rawRouteData.isIndex,
-    origin: rawRouteData.origin
-  };
-}
-
-function deserializeManifest(serializedManifest) {
-  const routes = [];
-  for (const serializedRoute of serializedManifest.routes) {
-    routes.push({
-      ...serializedRoute,
-      routeData: deserializeRouteData(serializedRoute.routeData)
-    });
-    const route = serializedRoute;
-    route.routeData = deserializeRouteData(serializedRoute.routeData);
-  }
-  const assets = new Set(serializedManifest.assets);
-  const componentMetadata = new Map(serializedManifest.componentMetadata);
-  const inlinedScripts = new Map(serializedManifest.inlinedScripts);
-  const clientDirectives = new Map(serializedManifest.clientDirectives);
-  const serverIslandNameMap = new Map(serializedManifest.serverIslandNameMap);
-  const key = decodeKey(serializedManifest.key);
-  return {
-    // in case user middleware exists, this no-op middleware will be reassigned (see plugin-ssr.ts)
-    middleware() {
-      return { onRequest: NOOP_MIDDLEWARE_FN };
-    },
-    ...serializedManifest,
-    assets,
-    componentMetadata,
-    inlinedScripts,
-    clientDirectives,
-    routes,
-    serverIslandNameMap,
-    key
-  };
-}
-
-const manifest = deserializeManifest({"hrefRoot":"file:///Users/user/Desktop/work/VLVU2025/","adapterName":"@astrojs/vercel","routes":[{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"page","component":"_server-islands.astro","params":["name"],"segments":[[{"content":"_server-islands","dynamic":false,"spread":false}],[{"content":"name","dynamic":true,"spread":false}]],"pattern":"^\\/_server-islands\\/([^/]+?)\\/?$","prerender":false,"isIndex":false,"fallbackRoutes":[],"route":"/_server-islands/[name]","origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"endpoint","isIndex":false,"route":"/_image","pattern":"^\\/_image\\/?$","segments":[[{"content":"_image","dynamic":false,"spread":false}]],"params":[],"component":"node_modules/.pnpm/astro@5.1.10_@types+node@22.10.7_jiti@1.21.7_rollup@4.31.0_typescript@5.7.3_yaml@2.7.0/node_modules/astro/dist/assets/endpoint/generic.js","pathname":"/_image","prerender":false,"fallbackRoutes":[],"origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"route":"/api/example","isIndex":true,"type":"endpoint","pattern":"^\\/api\\/example\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"example","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/example/index.ts","pathname":"/api/example","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"route":"/api/stamp","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/stamp\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"stamp","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/stamp.ts","pathname":"/api/stamp","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/game.DKnNW5Gf.css"}],"routeData":{"route":"/game","isIndex":false,"type":"page","pattern":"^\\/game\\/?$","segments":[[{"content":"game","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/game.astro","pathname":"/game","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/game.DKnNW5Gf.css"}],"routeData":{"route":"/","isIndex":true,"type":"page","pattern":"^\\/$","segments":[],"params":[],"component":"src/pages/index.astro","pathname":"/","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}}],"base":"/","trailingSlash":"ignore","compressHTML":true,"componentMetadata":[["/Users/user/Desktop/work/VLVU2025/src/pages/game.astro",{"propagation":"none","containsHead":true}],["/Users/user/Desktop/work/VLVU2025/src/pages/index.astro",{"propagation":"none","containsHead":true}]],"renderers":[],"clientDirectives":[["idle","(()=>{var l=(n,t)=>{let i=async()=>{await(await n())()},e=typeof t.value==\"object\"?t.value:void 0,s={timeout:e==null?void 0:e.timeout};\"requestIdleCallback\"in window?window.requestIdleCallback(i,s):setTimeout(i,s.timeout||200)};(self.Astro||(self.Astro={})).idle=l;window.dispatchEvent(new Event(\"astro:idle\"));})();"],["load","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).load=e;window.dispatchEvent(new Event(\"astro:load\"));})();"],["media","(()=>{var n=(a,t)=>{let i=async()=>{await(await a())()};if(t.value){let e=matchMedia(t.value);e.matches?i():e.addEventListener(\"change\",i,{once:!0})}};(self.Astro||(self.Astro={})).media=n;window.dispatchEvent(new Event(\"astro:media\"));})();"],["only","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).only=e;window.dispatchEvent(new Event(\"astro:only\"));})();"],["visible","(()=>{var a=(s,i,o)=>{let r=async()=>{await(await s())()},t=typeof i.value==\"object\"?i.value:void 0,c={rootMargin:t==null?void 0:t.rootMargin},n=new IntersectionObserver(e=>{for(let l of e)if(l.isIntersecting){n.disconnect(),r();break}},c);for(let e of o.children)n.observe(e)};(self.Astro||(self.Astro={})).visible=a;window.dispatchEvent(new Event(\"astro:visible\"));})();"]],"entryModules":{"\u0000@astrojs-ssr-adapter":"_@astrojs-ssr-adapter.mjs","\u0000noop-middleware":"_noop-middleware.mjs","\u0000@astro-page:src/pages/api/example/index@_@ts":"pages/api/example.astro.mjs","\u0000@astro-page:src/pages/api/stamp@_@ts":"pages/api/stamp.astro.mjs","\u0000@astro-page:src/pages/game@_@astro":"pages/game.astro.mjs","\u0000@astrojs-ssr-virtual-entry":"entry.mjs","\u0000@astro-page:src/pages/index@_@astro":"pages/index.astro.mjs","\u0000@astro-renderers":"renderers.mjs","\u0000@astro-page:node_modules/.pnpm/astro@5.1.10_@types+node@22.10.7_jiti@1.21.7_rollup@4.31.0_typescript@5.7.3_yaml@2.7.0/node_modules/astro/dist/assets/endpoint/generic@_@js":"pages/_image.astro.mjs","/Users/user/Desktop/work/VLVU2025/node_modules/.pnpm/astro@5.1.10_@types+node@22.10.7_jiti@1.21.7_rollup@4.31.0_typescript@5.7.3_yaml@2.7.0/node_modules/astro/dist/assets/services/sharp.js":"chunks/sharp_BYpgJSzI.mjs","\u0000@astrojs-manifest":"manifest_BP1gu5Kl.mjs","@astrojs/react/client.js":"_astro/client.BNC2RHkt.js","astro:scripts/before-hydration.js":""},"inlinedScripts":[],"assets":["/_astro/game.DKnNW5Gf.css","/favicon.svg","/_astro/client.BNC2RHkt.js","/fonts/AtsaneeDEMO-Thin.otf","/fonts/RibbonDemo-Regular.ttf","/fonts/YesevaOne-Regular.ttf"],"buildFormat":"directory","checkOrigin":true,"serverIslandNameMap":[],"key":"9mojHcSash8k7bsQ4+Y+lkbe9gPxoJgjJBnYvwEivkI="});
-if (manifest.sessionConfig) manifest.sessionConfig.driverModule = null;
-
-export { manifest };
